Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -6,8 +6,8 @@
 
 int main() {
     // printf("%s", welcome());
-    float a;
-    scanf("%.4f", &a);
-    printf("You entered: %.4f", a);
+    double a;
+    scanf("%d", &a);
+    printf("You entered: %.4d", a);
     return 0;
 }
